// apps/frontend/src/api/generated/model/mappingDecision.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * WDIAG — Where Did It All Go
 * Personal finance budgeting and intelligence
 * OpenAPI spec version: 0.1.0
 */
import type { MappingDecisionAction } from './mappingDecisionAction'
import type { MappingDecisionSystemAccountId } from './mappingDecisionSystemAccountId'
import type { MappingDecisionNewAccount } from './mappingDecisionNewAccount'

export interface MappingDecision {
  simplefin_account_id: string
  action: MappingDecisionAction
  system_account_id?: MappingDecisionSystemAccountId
  authoritative?: boolean
  new_account?: MappingDecisionNewAccount
}